+++ app.mjs
@@ -1,4 +1,3 @@
-import fs from 'fs';
 import { NODE_ENV, PORT, SCROBBLERBOT_TOKEN } from './config';
 import { sendToAdmin } from './src/helpers/util';
 import bot from './src/bot';
@@ -7,12 +6,7 @@ import bot from './src/bot';
 if (NODE_ENV === 'production') {
   bot.startWebhook(`/${SCROBBLERBOT_TOKEN}`, null, PORT);
 } else {
-  const tlsOptions = {
-    key: fs.readFileSync('key.key'),
-    cert: fs.readFileSync('cert.pem'),
-  };
-
-  bot.startWebhook(`/${SCROBBLERBOT_TOKEN}`, tlsOptions, PORT);
+  bot.startPolling();
 }
 
 console.log(`Server is listening to localhost:${PORT}`);
