+++ src/bot.mjs
@@ -1,6 +1,7 @@
 import Telegraf from 'telegraf';
-import { scrobbleTracks, scrobbleTrackFromText, scrobbleTracksByParts } from './helpers/scrobbler';
-import { searchFromLastfmAndAnswerInlineQuery } from './helpers/actions';
+import HttpsProxyAgent from 'https-proxy-agent';
+import { scrobbleTrackFromText } from './helpers/scrobbler';
+import { searchFromLastfmAndAnswerInlineQuery, cancel, retry, repeat, repeatMany } from './handlers/actions';
 import user from './middlewares/user';
 import scenes from './middlewares/scenes';
 import auth from './middlewares/auth';
@@ -10,8 +11,7 @@ import logger from './middlewares/logger';
 import error from './middlewares/error';
 import asyncer from './middlewares/asyncer';
 import { SCROBBLERBOT_TOKEN } from '../config';
-import { successfulScrobble, scrobbleError, multipleArray, sendToAdmin } from './helpers/util';
-import { findSucceededMessageById, findFailedMessageById } from './helpers/dbmanager';
+import { sendToAdmin } from './helpers/util';
 import './helpers/scheduler';
 import './db';
 
@@ -19,6 +19,11 @@ import './db';
 const bot = new Telegraf(SCROBBLERBOT_TOKEN, {
   telegram: {
     webhookReply: false,
+    agent: process.env.NODE_ENV === 'development' ? new HttpsProxyAgent({ // proxy for russian blocked devs
+      host: '207.154.197.214',
+      port: 8888,
+      secureProxy: true,
+    }) : null,
   },
 });
 
@@ -29,7 +34,7 @@ setImmediate(async () => {
     const data = await bot.telegram.getMe();
     bot.options.username = data.username;
   } catch (e) {
-    console.log(e.message);
+    console.error(e);
   }
 });
 
@@ -44,94 +49,19 @@ bot.hears(/\/\w+/, async (ctx) => {
   await ctx.reply('If you are confused type /help');
 });
 
-bot.on('text', auth, limiter, async (ctx) => {
-  await scrobbleTrackFromText(ctx);
-});
+bot.on('text', auth, limiter, scrobbleTrackFromText);
 
 bot.on('inline_query', async (ctx) => {
   await searchFromLastfmAndAnswerInlineQuery(ctx, 'track');
 });
 
-bot.action('CANCEL', async (ctx) => {
-  await ctx.flow.leave();
-  await ctx.editMessageText('Canceled');
-});
-
-bot.action('RETRY', limiter, async (ctx) => {
-  ctx.flow.state.messageIdToEdit = (await ctx.editMessageText('<i>Scrobbling...</i>',
-    Telegraf.Extra.HTML())).message_id;
-  const messageId = ctx.callbackQuery.message.message_id;
-  const message = await findFailedMessageById(messageId);
-
-  if (!message) {
-    await ctx.editMessageText('Expired');
-    return;
-  }
-
-  try {
-    await scrobbleTracks(message.tracks, undefined, ctx.user.key);
-  } catch (e) {
-    await scrobbleError(ctx, e, message.tracks);
-    return;
-  }
-
-  await successfulScrobble(ctx);
-});
-
-bot.action('REPEAT', async (ctx) => {
-  const messageId = ctx.callbackQuery.message.message_id;
-  const message = await findSucceededMessageById(messageId);
-
-  if (!message) {
-    await ctx.editMessageText('Expired');
-    return;
-  }
-
-  await ctx.editMessageText('How many times do you want to scrobble this again?',
-    Telegraf.Markup.inlineKeyboard([
-      [
-        Telegraf.Markup.callbackButton('1', 'REPEAT:1'),
-        Telegraf.Markup.callbackButton('2', 'REPEAT:2'),
-        Telegraf.Markup.callbackButton('3', 'REPEAT:3'),
-        Telegraf.Markup.callbackButton('4', 'REPEAT:4'),
-        Telegraf.Markup.callbackButton('5', 'REPEAT:5'),
-      ],
-      [
-        Telegraf.Markup.callbackButton('6', 'REPEAT:6'),
-        Telegraf.Markup.callbackButton('7', 'REPEAT:7'),
-        Telegraf.Markup.callbackButton('8', 'REPEAT:8'),
-        Telegraf.Markup.callbackButton('9', 'REPEAT:9'),
-        Telegraf.Markup.callbackButton('10', 'REPEAT:10'),
-      ],
-    ]).extra());
-});
-
-bot.action(/REPEAT:\d?\d/, limiter, async (ctx) => {
-  ctx.flow.state.messageIdToEdit = (await ctx.editMessageText('<i>Scrobbling...</i>',
-    Telegraf.Extra.HTML())).message_id;
-  const messageId = ctx.callbackQuery.message.message_id;
-  const message = await findSucceededMessageById(messageId);
-
-  if (!message) {
-    await ctx.editMessageText('Expired');
-    return;
-  }
-
-  const count = ctx.callbackQuery.data.split(':')[1];
-  const tracks = multipleArray(message.tracks, count);
-
-  try {
-    await scrobbleTracksByParts(ctx, tracks);
-  } catch (e) {
-    await scrobbleError(ctx, e, tracks);
-    return;
-  }
-
-  await successfulScrobble(ctx, undefined, message.tracks);
-});
+bot.action('CANCEL', cancel);
+bot.action('RETRY', limiter, retry);
+bot.action('REPEAT', repeat);
+bot.action(/REPEAT:\d?\d/, limiter, repeatMany);
 
 bot.catch((e) => {
-  console.log(e);
+  console.error(e);
   sendToAdmin(`Unhandled Bot Error! ${e.message}`);
 });
